dossier/requirements.txt
########################################
# DEFAULT CONFIGURATION
########################################
# GENERAL
openpyxl==3.0.6
pyaml==20.4.0
pyarrow==2.0.0

# INFRA
cx-Oracle==8.1.0
SQLAlchemy==1.3.23
pysftp==0.2.9

# DATA / IA
numpy==1.19.5
scipy==1.5.4
pandas==1.1.5
xlrd==2.0.1
pyspark==3.2.1
scikit-learn==0.24.1

# VISUALISATION
matplotlib==3.3.4
pandas-profiling==3.0.0

# DOC
Sphinx==3.4.3
sphinx-rtd-theme==0.5.1

# TEST
pytest==6.2.1
pytest-spark==0.6.0
pytest-cov==2.11.1
coverage==5.3.1

# FORMAT
black==19.3b0

# AZURE
azureml-core==1.20.0
azureml-train==1.20.0
azureml-widgets==1.20.0

########################################
# PROJECT SPECIFIC ADD-ONS
########################################
